@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.5.3
+  sdk: ^3.6.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -37,13 +37,13 @@ dependencies:
   cupertino_icons: ^1.0.8
   get_it: ^8.0.1
   go_router: ^14.3.0
-  internet_connection_checker: ^1.0.0+1
+  internet_connection_checker: ^3.0.1
   equatable: ^2.0.5
   flutter_bloc: ^8.1.6
   flutter_screenutil: ^5.9.3
   font_awesome_flutter: ^10.7.0
   pin_code_fields: ^8.0.1
-  fl_chart: ^0.69.0
+  fl_chart: ^0.70.0
   flutter_native_splash: ^2.4.2
   firebase_auth: ^5.3.1
   cloud_firestore: ^5.4.4
@@ -66,13 +66,39 @@ dependencies:
   dropdown_button2: ^2.3.9
   supabase_flutter: ^2.8.0
   build_runner: ^2.4.13
+  easy_localization: ^3.0.7
+  svg_flutter: ^0.0.1
+  flutter_gen: ^5.8.0
+  percent_indicator: ^4.2.4
+  toastification: ^2.3.0
+  geekyants_flutter_gauges: ^1.0.4
+  flutter_launcher_icons: ^0.14.3
+  flutter_secure_storage: ^9.2.4
+
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
+  device_info_plus: ^11.2.2
+  battery_plus: ^6.2.1
+  hive_generator: ^2.0.1
+  syncfusion_flutter_gauges: ^28.2.3
   # shared_preferences: ^2.3.2
 
-# flutter_native_splash:
-#   color: "#071b5d"
-#   image: assets/images/clot2.png
-#   android: true
-#   ios: true
+flutter_native_splash:
+  color: "#071b5d"
+  image: assets/icons/splash_logo.png
+  android: true
+  # android_12:
+  #   # image: assets/icons/splashLogo.png
+  #   # color: "#071b5d"
+  ios: true
+flutter_launcher_icons:
+  android: true
+  ios: true
+  adaptive_icon_background: "#071b5d"
+  image_background_color: "#071b5d"
+  adaptive_icon_foreground: "assets/icons/splash_logo.png"
+  image_path: "assets/icons/splash_logo.png"
+  min_sdk_android: 21 # android min sdk min:16, default 21
 
 dev_dependencies:
   flutter_test:
@@ -83,7 +109,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^4.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -99,6 +125,8 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
       - assets/images/
+      - assets/icons/
+      - assets/translation/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
