@@ -1,4 +1,4 @@
-library config;
+library;
 
 export 'injection.dart';
 export 'keyboard.dart';
