@@ -1,4 +1,4 @@
-library app_colors;
+library;
 
 import 'package:flutter/material.dart';
 
