@@ -1,9 +1,12 @@
 class Constant {
-  static const String baseUrl =
-      'https://api.mockfly.dev/mocks/be49e904-cb4c-467c-b5f9-097e6ccb1e17';
+  static const String baseUrl = 'https://siaray.holylands.net/api';
   static const String loginEndpoint = '/login';
   static const String token = '';
   static const supabaseUrl = 'https://kehnastnwntyzztjhijd.supabase.co';
   static const supabaseKey =
       'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImtlaG5hc3Rud250eXp6dGpoaWpkIiwicm9sZSI6ImFub24iLCJpYXQiOjE3MzEwMTA3ODksImV4cCI6MjA0NjU4Njc4OX0.qD__s9vDnRtxsxm5Uxbgx9-LWphPaz6Fp0htSWY036w';
+  static const String forgotPasswordEndpoint = '/user/auth/forgot-password';
+  static const String otpVerificationEndpoint = '/user/auth/verify-reset-code';
+  static const String resetPasswordEndpoint = '/user/auth/reset-password';
+  static const String getBatteries = '/user/battery';
 }
