@@ -1,4 +1,4 @@
-library constants;
+library;
 
 export 'colors.dart';
 export 'icons.dart';
