@@ -1,1 +1,1 @@
-library global;
+library;
