@@ -1,1 +1,1 @@
-library localization;
+library;
