@@ -1,4 +1,4 @@
-library usecases;
+library;
 
 abstract class Usecases<type, params> {
   Future<type> call({params? params});
