@@ -1,10 +1,8 @@
- 
-    class  DevicesModel {
-        // ...
-        // A model is a more generic term and can represent structured data used within an application. 
-        // ...
-        // fromJson
-        // ...
-        // toJson
-   
-    }
+class DevicesModel {
+  // ...
+  // A model is a more generic term and can represent structured data used within an application.
+  // ...
+  // fromJson
+  // ...
+  // toJson
+}
