@@ -1,6 +1,4 @@
-
-    class DevicesEntity {
-        // ...
-        // An entity represents a real-world object with a distinct identity. 
-    }
-    
+class DevicesEntity {
+  // ...
+  // An entity represents a real-world object with a distinct identity.
+}
