@@ -1,13 +1,11 @@
- 
-    import '../repositories/repositories.dart';
+import '../repositories/repositories.dart';
 
-    class GetDevicesUseCase {
-        final DevicesRepository repository;
-      
-        GetDevicesUseCase({required this.repository});
-      
-        // Future<User> execute(String userId) async {
-        //   return userRepository.getUser(userId);
-        // }
-      }
-      
+class GetDevicesUseCase {
+  final DevicesRepository repository;
+
+  GetDevicesUseCase({required this.repository});
+
+  // Future<User> execute(String userId) async {
+  //   return userRepository.getUser(userId);
+  // }
+}
