@@ -1,10 +1,8 @@
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:equatable/equatable.dart';
 
-  import 'package:flutter_bloc/flutter_bloc.dart';
-  import 'package:equatable/equatable.dart';
-  
-  part 'devices_state.dart';
-  
-  class DevicesCubit extends Cubit<DevicesState> {
-    DevicesCubit() : super(DevicesInitial());
-  }
-  
+part 'devices_state.dart';
+
+class DevicesCubit extends Cubit<DevicesState> {
+  DevicesCubit() : super(DevicesInitial());
+}
