@@ -1,11 +1,10 @@
+part of 'devices_cubit.dart';
 
-  part of 'devices_cubit.dart';
-  
-  abstract class DevicesState extends Equatable {
+abstract class DevicesState extends Equatable {
   const DevicesState();
 
   @override
   List<Object> get props => [];
-  }
-  class DevicesInitial extends DevicesState {}
-  
+}
+
+class DevicesInitial extends DevicesState {}
